+++ src/components/pages/entities/Entities.jsx
@@ -623,7 +623,7 @@ function Entities() {
                               permissionCheck={(permission) =>
                                 permission.module === 'entity' && permission.canRead
                               }>
-                              <Link to={`${row.id}`}>
+                              <Link to={`${row.id}/overview`}>
                                 <p className='truncate text-xs'> {value}</p>
                               </Link>
                             </GateKeeper>
@@ -678,7 +678,7 @@ function Entities() {
                           <button
                             type='button'
                             className=' inline-flex items-center gap-x-1 text-sm font-semibold rounded-lg  text-[#475569] hover:text-orange-500 disabled:opacity-50 disabled:pointer-events-none dark:focus:outline-none dark:focus:ring-1 dark:focus:ring-gray-600'>
-                            <Link to={`${row.id}`}>
+                            <Link to={`${row.id}/overview`}>
                               <svg
                                 xmlns='http://www.w3.org/2000/svg'
                                 viewBox='0 0 20 20'
