+++ src/routerLayer/entity/entity.router.js
@@ -1,6 +1,10 @@
 import EntityForm, { entityFormLoader } from "../../components/createForm/createEntityForm/EntityForm";
+import Boardmeeting from "../../components/landingPages/ReuseableComponents/Boardmeeting";
+import Documents from "../../components/landingPages/ReuseableComponents/Documents";
 import EntityLandingPage from "../../components/landingPages/entity/EntityLandingPage";
 import Overview from "../../components/landingPages/entity/Overview";
+import Task from "../../components/landingPages/ReuseableComponents/Task";
+import BoardMeetings from "../../components/pages/boardMeetings/BoardMeetings";
 import Entities, { action as entityAction, loader as entityLoader } from "../../components/pages/entities/Entities";
 import RouteBlocker from "../../rbac/RouteBlocker";
 
@@ -11,16 +15,10 @@ export const entityRouter = [
         children: [
             { index: true, loader: entityLoader, action: entityAction, element: <Entities />, },
             { path: ':id', element: <EntityLandingPage />, children: [
-                { path: 'overview', element: <Overview /> },
-                { path: 'task', element: <Overview /> },
-
-                { path: 'boardmeetings', element: <Overview /> },
-
-                { path: 'documents', element: <Overview /> },
-
-
-                // { path: 'sub-route-2', element: <SubComponent2 /> }
-                // Add more sub-children as needed
+                { path: 'overview', element: <Overview />  },
+                { path: 'task', element: <Task /> },
+                { path: 'boardmeetings', element: <Boardmeeting /> },
+                { path: 'documents', element: <Documents/> },
             ] },
         ]
     },
